Colour scheme should not allow other choices.
Currently, if you enter a random string in the `palette` property and
press enter, the template will save the value, then throw an error.
This is obviously undesirable behaviour and it would be better if
the template validated the colour value.

In addition, the existing `custom_colors` property already allows
you to set a custom set of hex values or strings as your colour scheme.

In time we should get rid of `custom_colors`, add more validation in the
template, and re-enable `choices_other` on `palette`: but for the moment,
this is a simple fix.